// src/Models/coffee.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0' at 'Thu, 07 Apr 2022 08:01:06 GMT'
 */
export type Coffee = IContentItem<{
  metadataOgDescription: Elements.TextElement;
  metadataMetaTitle: Elements.TextElement;
  metadataOgTitle: Elements.TextElement;
  productStatus: Elements.TaxonomyElement;
  altitude: Elements.TextElement;
  metadataMetaDescription: Elements.TextElement;
  variety: Elements.TextElement;
  image: Elements.AssetsElement;
  metadataTwitterSite: Elements.TextElement;
  urlPattern: Elements.UrlSlugElement;
  price: Elements.NumberElement;
  metadataTwitterImage: Elements.AssetsElement;
  metadataTwitterCreator: Elements.TextElement;
  country: Elements.TextElement;
  sitemap: Elements.TaxonomyElement;
  metadataTwitterTitle: Elements.TextElement;
  shortDescription: Elements.RichTextElement;
  processing: Elements.TaxonomyElement;
  metadataTwitterDescription: Elements.TextElement;
  metadataOgImage: Elements.AssetsElement;
  longDescription: Elements.RichTextElement;
  farm: Elements.TextElement;
  productName: Elements.TextElement;
}>;